Multiple refactors
Refactors:
 * Move transport package to comm
 * Use Linked instead of Array BlockingQueue
 * Move protobuf generated files to Protocol class
