+++ app/src/androidTest/java/programatorus/mobile/transport/TestUtils.kt
@@ -1,12 +1,11 @@
 package programatorus.mobile.transport
 
-import programus.proto.GenericMessage
-import programus.proto.TestMessage
+import programus.proto.Protocol
 
 object TestUtils {
-    fun newTestMessage() = GenericMessage.newBuilder()
+    fun newTestMessage() = Protocol.GenericMessage.newBuilder()
         .setSessionId(10)
-        .setTest(TestMessage.newBuilder().apply {
+        .setTest(Protocol.TestMessage.newBuilder().apply {
             value = "Test 1234"
         })
         .build()
