+++ app/src/main/java/programatorus/client/comm/transport/IOutgoingPacket.kt
@@ -1,4 +1,4 @@
-package programatorus.client.transport
+package programatorus.client.comm.transport
 
 import java.util.concurrent.CompletableFuture
