+++ app/src/main/java/programatorus/client/comm/transport/ITransport.kt
@@ -1,4 +1,4 @@
-package programatorus.client.transport
+package programatorus.client.comm.transport
 
 interface ITransport {
