+++ app/src/main/java/programatorus/client/comm/transport/io/PipedTransport.kt
@@ -1,8 +1,8 @@
-package programatorus.client.transport.io
+package programatorus.client.comm.transport.io
 
 import android.os.Handler
 import android.os.Looper
-import programatorus.client.transport.ITransportClient
+import programatorus.client.comm.transport.ITransportClient
 import java.io.InputStream
 import java.io.OutputStream
 import java.io.PipedInputStream
@@ -11,10 +11,10 @@ import java.io.PipedOutputStream
 /**
  * Mostly for testing purposes
  */
-class LoopbackTransport(
+class PipedTransport(
     client: ITransportClient,
     executor: Handler = Handler(Looper.getMainLooper())
-) : IOTransport<LoopbackTransport>(client, executor) {
+) : StreamingTransport<PipedTransport>(client, executor) {
 
     override var inputStream: InputStream? = null
